requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
discord.py==1.7.3
emoji==1.2.0
emojis==0.6.0
idna==2.10
multidict==5.1.0
nanoid==2.0.0
numpy==1.21.1
opencv-python==4.5.3.56
Pillow==8.3.1
pycoingecko==2.2.0
python-dotenv==0.18.0
requests==2.25.1
tinydb==4.5.0
tornado==6.1
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
